# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VendorBridge Procurement & Vendor Management ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already in use
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent

  # Users
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List approvals
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
  /approvals/{id}/approve:
    post:
      operationId: approveApproval
      tags: [approvals]
      summary: Approve a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionInput"
      responses:
        "200":
          description: Approval approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{id}/reject:
    post:
      operationId: rejectApproval
      tags: [approvals]
      summary: Reject a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionInput"
      responses:
        "200":
          description: Approval rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"

  # Purchase Orders
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice from purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/send-email:
    post:
      operationId: sendInvoiceEmail
      tags: [invoices]
      summary: Send invoice via email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailInput"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required: [message]

  # Activity Logs
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    SignupInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [procurement_officer, vendor, manager, admin]
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # User schemas
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [procurement_officer, vendor, manager, admin]
        isActive:
          type: boolean
        vendorId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Approval:
      type: object
      required: [id, quotationId, status, createdAt]
      properties:
        id:
          type: integer
        quotationId:
          type: integer
        rfqTitle:
          type: ["string", "null"]
        vendorName:
          type: ["string", "null"]
        totalPrice:
          type: ["number", "null"]
        approvedById:
          type: ["integer", "null"]
        approvedByName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ApprovalActionInput:
      type: object
      properties:
        remarks:
          type: string

    # Purchase Order schemas
...
    Invoice:
      type: object
      required: [id, invoiceNumber, purchaseOrderId, vendorId, subtotal, taxAmount, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        purchaseOrderId:
          type: integer
        poNumber:
          type: ["string", "null"]
        vendorId:
          type: integer
        vendorName:
          type: ["string", "null"]
        subtotal:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        emailSentAt:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceInput:
      type: object
      required: [purchaseOrderId, vendorId, subtotal, taxAmount, totalAmount]
      properties:
        purchaseOrderId:
          type: integer
        vendorId:
          type: integer
        subtotal:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        dueDate:
          type: string
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
        notes:
          type: string
    SendEmailInput:
      type: object
      required: [recipientEmail]
      properties:
        recipientEmail:
          type: string
        subject:
          type: string
        message:
          type: string

    # Activity Log schemas
...
    Notification:
      type: object
      required: [id, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [rfq, quotation, approval, po, invoice, general]
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string

    # Dashboard schemas
...