# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VendorBridge Procurement & Vendor Management ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStatus } from './approvalStatus';

export interface Approval {
  id: number;
  quotationId: number;
  /** @nullable */
  rfqTitle?: string | null;
  /** @nullable */
  vendorName?: string | null;
  /** @nullable */
  totalPrice?: number | null;
  /** @nullable */
  approvedById?: number | null;
  /** @nullable */
  approvedByName?: string | null;
  status: ApprovalStatus;
  /** @nullable */
  remarks?: string | null;
  createdAt: string;
  updatedAt?: string;
}